Various changes related to 68000 interrupts (nw) - Use 8-bit handlers for 68000 vector fetches - mc68307: Use internal map for interrupt acknowledgment (was overlooked) - tti: Change interrupt level from 5 to 2 - mc68901: Add (untested) callback for daisy-chained interrupt acknowledgment

+++ src/devices/machine/68340.h
@@ -57,7 +57,7 @@ private:
 
 	void update_ipl();
 	void internal_vectors_r(address_map &map);
-	u16 int_ack(offs_t offset);
+	uint8_t int_ack(offs_t offset);
 
 	TIMER_CALLBACK_MEMBER(periodic_interrupt_timer_callback);
